requirements.txt
streamlit
pandas
google-generativeai
matplotlib
seaborn
python-dotenv
scikit-learn 